refactor: Move handlers to separate file
