+++ internal/server/server.go
@@ -1,16 +1,12 @@
 package server
 
 import (
-	"encoding/json"
 	"fmt"
 	"io/fs"
 	"log/slog"
 	"net/http"
 	"os"
-	"path/filepath"
-	"strings"
 
-	"github.com/google/uuid"
 	"github.com/gorilla/websocket"
 	"github.com/terrabitz/rpg-audio-streamer/internal/auth"
 	"github.com/terrabitz/rpg-audio-streamer/internal/middlewares"
@@ -50,7 +46,6 @@ type Config struct {
 }
 
 func New(cfg Config, logger *slog.Logger, auth Authenticator, store Store, hub WSRegisterer) (*Server, error) {
-
 	srv := &Server{
 		logger: logger,
 		cfg:    cfg,
@@ -75,20 +70,7 @@ func (s *Server) Start() error {
 		return fmt.Errorf("failed to create upload directory: %w", err)
 	}
 
-	mux := http.NewServeMux()
-
-	// Public endpoints
-	mux.HandleFunc("/api/v1/login", s.handleLogin)
-	mux.HandleFunc("/api/v1/join", s.handleJoin)
-	mux.HandleFunc("/api/v1/auth/status", s.handleAuthStatus)
-	mux.HandleFunc("/api/v1/auth/logout", s.handleLogout)
-
-	// Protected endpoints with role validation
-	mux.HandleFunc("/api/v1/files", s.gmOnlyMiddleware(s.handleFiles))
-	mux.HandleFunc("/api/v1/files/{trackID}", s.gmOnlyMiddleware(s.handleFileDelete))
-	mux.HandleFunc("/api/v1/join-token", s.gmOnlyMiddleware(s.handleJoinToken))
-	mux.HandleFunc("/api/v1/stream/", s.authMiddleware(s.streamDirectory))
-	mux.HandleFunc("/api/v1/trackTypes", s.authMiddleware(s.handleTrackTypes))
+	mux := s.registerHandlers()
 
 	// Apply global middleware
 	handler := middlewares.LoggerMiddleware(s.logger)(
@@ -114,207 +96,21 @@ func (s *Server) Start() error {
 	return srv.ListenAndServe()
 }
 
-func (s *Server) handleFiles(w http.ResponseWriter, r *http.Request, token *auth.Token) {
-	switch r.Method {
-	case http.MethodGet:
-		s.listFiles(w, r)
-	case http.MethodPost:
-		s.uploadFile(w, r)
-	default:
-		http.Error(w, "Method not allowed", http.StatusMethodNotAllowed)
-	}
-}
+func (s *Server) registerHandlers() *http.ServeMux {
+	mux := http.NewServeMux()
 
-func (s *Server) handleFileDelete(w http.ResponseWriter, r *http.Request, token *auth.Token) {
-	if r.Method != http.MethodDelete {
-		http.Error(w, "Method not allowed", http.StatusMethodNotAllowed)
-		return
-	}
+	// Public endpoints
+	mux.HandleFunc("/api/v1/login", s.handleLogin)
+	mux.HandleFunc("/api/v1/join", s.handleJoin)
+	mux.HandleFunc("/api/v1/auth/status", s.handleAuthStatus)
+	mux.HandleFunc("/api/v1/auth/logout", s.handleLogout)
 
-	trackIDString := r.PathValue("trackID")
-	trackID, err := uuid.Parse(trackIDString)
-	if err != nil {
-		http.Error(w, "Invalid track ID", http.StatusBadRequest)
-		return
-	}
+	// Protected endpoints with role validation
+	mux.HandleFunc("/api/v1/files", s.gmOnlyMiddleware(s.handleFiles))
+	mux.HandleFunc("/api/v1/files/{trackID}", s.gmOnlyMiddleware(s.handleFileDelete))
+	mux.HandleFunc("/api/v1/join-token", s.gmOnlyMiddleware(s.handleJoinToken))
+	mux.HandleFunc("/api/v1/stream/", s.authMiddleware(s.streamDirectory))
+	mux.HandleFunc("/api/v1/trackTypes", s.authMiddleware(s.handleTrackTypes))
 
-	// Retrieve the track to get its folder path
-	track, err := s.store.GetTrackByID(r.Context(), trackID)
-	if err != nil {
-		http.Error(w, "Track not found", http.StatusNotFound)
-		return
-	}
-
-	// Remove the database record
-	if err := s.store.DeleteTrack(r.Context(), trackID); err != nil {
-		http.Error(w, "Failed to remove track record", http.StatusInternalServerError)
-		return
-	}
-
-	if err := os.RemoveAll(track.Path); err != nil {
-		http.Error(w, "Failed to delete folder", http.StatusInternalServerError)
-		return
-	}
-
-	w.WriteHeader(http.StatusOK)
-	w.Write([]byte("Track deleted successfully"))
-}
-
-func (s *Server) streamDirectory(w http.ResponseWriter, r *http.Request, token *auth.Token) {
-	relativePath := strings.TrimPrefix(r.URL.Path, "/api/v1/stream/")
-	filePath := filepath.Join(s.cfg.UploadDir, relativePath)
-	http.ServeFile(w, r, filePath)
-}
-
-func (s *Server) handleWebSocket(w http.ResponseWriter, r *http.Request, token *auth.Token) {
-	conn, err := s.upgrader.Upgrade(w, r, nil)
-	if err != nil {
-		s.logger.Error("websocket upgrade failed", "error", err)
-		return
-	}
-
-	s.hub.Register(conn, token)
-}
-
-type loginRequest struct {
-	Username string `json:"username"`
-	Password string `json:"password"`
-}
-
-type loginResponse struct {
-	Success bool   `json:"success"`
-	Error   string `json:"error,omitempty"`
-}
-
-type authStatusResponse struct {
-	Authenticated bool      `json:"authenticated"`
-	Role          auth.Role `json:"role,omitempty"`
-}
-
-type joinTokenResponse struct {
-	Token string `json:"token"`
-}
-
-type joinRequest struct {
-	Token string `json:"token"`
-}
-
-func (s *Server) handleAuthStatus(w http.ResponseWriter, r *http.Request) {
-	if r.Method != http.MethodGet {
-		http.Error(w, "Method not allowed", http.StatusMethodNotAllowed)
-		return
-	}
-
-	cookie, err := readCookie(r, authCookieName)
-	if err != nil {
-		w.Header().Set("Content-Type", "application/json")
-		json.NewEncoder(w).Encode(authStatusResponse{Authenticated: false})
-		return
-	}
-
-	token, err := s.auth.ValidateToken(cookie)
-	w.Header().Set("Content-Type", "application/json")
-	if err != nil {
-		json.NewEncoder(w).Encode(authStatusResponse{Authenticated: false})
-		return
-	}
-
-	json.NewEncoder(w).Encode(authStatusResponse{
-		Authenticated: true,
-		Role:          token.Role,
-	})
-}
-
-func (s *Server) handleLogout(w http.ResponseWriter, r *http.Request) {
-	if r.Method != http.MethodPost {
-		http.Error(w, "Method not allowed", http.StatusMethodNotAllowed)
-		return
-	}
-
-	s.clearCookie(w, authCookieName)
-	w.WriteHeader(http.StatusOK)
-}
-
-func (s *Server) handleLogin(w http.ResponseWriter, r *http.Request) {
-	if r.Method != http.MethodPost {
-		http.Error(w, "Method not allowed", http.StatusMethodNotAllowed)
-		return
-	}
-
-	var req loginRequest
-	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
-		s.logger.Error("failed to decode login request", "error", err)
-		http.Error(w, "Invalid request body", http.StatusBadRequest)
-		return
-	}
-
-	creds := auth.Credentials{
-		Username: req.Username,
-		Password: req.Password,
-	}
-
-	token, err := s.auth.ValidateCredentials(creds)
-	if err != nil {
-		s.logger.Info("login failed", "username", req.Username)
-		w.WriteHeader(http.StatusUnauthorized)
-		json.NewEncoder(w).Encode(loginResponse{
-			Success: false,
-			Error:   "Invalid credentials",
-		})
-		return
-	}
-
-	// Set auth cookie
-	s.writeCookie(w, authCookieName, token.String(), token.ExpiresAt)
-
-	w.Header().Set("Content-Type", "application/json")
-	json.NewEncoder(w).Encode(loginResponse{
-		Success: true,
-	})
-}
-
-func (s *Server) handleJoinToken(w http.ResponseWriter, r *http.Request, token *auth.Token) {
-	if r.Method != http.MethodGet {
-		http.Error(w, "Method not allowed", http.StatusMethodNotAllowed)
-		return
-	}
-
-	resp := joinTokenResponse{
-		Token: s.auth.GetJoinToken(),
-	}
-
-	w.Header().Set("Content-Type", "application/json")
-	json.NewEncoder(w).Encode(resp)
-}
-
-func (s *Server) handleJoin(w http.ResponseWriter, r *http.Request) {
-	if r.Method != http.MethodPost {
-		http.Error(w, "Method not allowed", http.StatusMethodNotAllowed)
-		return
-	}
-
-	var req joinRequest
-	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
-		s.logger.Error("failed to decode join request", "error", err)
-		http.Error(w, "Invalid request body", http.StatusBadRequest)
-		return
-	}
-
-	token, err := s.auth.ValidateJoinToken(req.Token)
-	if err != nil {
-		s.logger.Info("join failed", "error", err)
-		w.WriteHeader(http.StatusUnauthorized)
-		json.NewEncoder(w).Encode(loginResponse{
-			Success: false,
-			Error:   "Invalid join token",
-		})
-		return
-	}
-
-	s.writeCookie(w, authCookieName, token.String(), token.ExpiresAt)
-
-	w.Header().Set("Content-Type", "application/json")
-	json.NewEncoder(w).Encode(loginResponse{
-		Success: true,
-	})
+	return mux
 }
